+++ src/gt4py/next/program_processors/runners/dace_fieldview/utility.py
@@ -16,10 +16,24 @@ from typing import Any, Mapping
 
 import dace
 
+from gt4py.eve import codegen
+from gt4py.eve.codegen import FormatTemplate as as_fmt
 from gt4py.next.common import Connectivity, Dimension
 from gt4py.next.iterator import ir as itir
 from gt4py.next.iterator.ir_utils import common_pattern_matcher as cpm
 from gt4py.next.type_system import type_specifications as ts
+
+
+class SymbolicTranslator(codegen.TemplatedGenerator):
+    SymRef = as_fmt("{id}")
+    Literal = as_fmt("{value}")
+
+    def visit_FunCall(self, node: itir.FunCall) -> str:
+        if cpm.is_call_to(node, "deref"):
+            assert len(node.args) == 1
+            return self.visit(node.args[0])
+
+        raise RuntimeError(f"Unexpected 'FunCall' expression ({node}).")
 
 
 def as_dace_type(type_: ts.ScalarType) -> dace.typeclass:
@@ -78,6 +92,12 @@ def get_domain(
         domain[dim] = (bounds[0], bounds[1])
 
     return domain
+
+
+def get_symbolic_expr(
+    node: itir.Expr,
+) -> str:
+    return SymbolicTranslator().visit(node)
 
 
 def unique_name(prefix: str) -> str:
